refactor: update extendCommonOrder to use api-utils fn
