+++ package.json
@@ -25,7 +25,7 @@
   },
   "sideEffects": false,
   "dependencies": {
-    "@reactioncommerce/api-utils": "^1.9.0",
+    "@reactioncommerce/api-utils": "^1.12.0",
     "@reactioncommerce/logger": "^1.1.3",
     "@reactioncommerce/reaction-error": "^1.0.1",
     "lodash": "^4.17.15",
